pyproject.toml
[project]
name = "aduanas-conecta-logis-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrés Wallberg",email = "awallberg@entel.cl"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "duckdb (>=1.3.1,<2.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "prefect (>=3.0)",
    "numpy (>=2.3.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
